Moved service code to controller code, more fitting
